fix extraManipulation
